+++ src/components/Cell.tsx
@@ -4,12 +4,14 @@ import cx from 'classnames'
 export const Cell: FC<{
   gutter: boolean
   stickyRight: boolean
+  stickyLeft?: number
   disabled?: boolean
   className?: string
   active?: boolean
   children?: any
   width: number
-  left: number
+  left?: number
+  positionRelative?: boolean
 }> = ({
   children,
   gutter,
@@ -19,6 +21,8 @@ export const Cell: FC<{
   className,
   width,
   left,
+  stickyLeft,
+  positionRelative,
 }) => {
   return (
     <div
@@ -28,11 +32,13 @@ export const Cell: FC<{
         disabled && 'dsg-cell-disabled',
         gutter && active && 'dsg-cell-gutter-active',
         stickyRight && 'dsg-cell-sticky-right',
+        stickyLeft !== undefined && 'dsg-cell-sticky-left',
         className
       )}
       style={{
         width,
-        left: stickyRight ? undefined : left,
+        left: stickyLeft !== undefined ? stickyLeft : (stickyRight ? undefined : left),
+        ...(positionRelative ? { position: 'relative' } : {}),
       }}
     >
       {children}