+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.13.1
+> Date: 2025-07-16
+- Fixed frozenColumns for multiple columns use case
+
+## 4.13.0
+> Date: 2025-07-16
+- Added frozenColumns property on DataSheetGrid component to freeze X first column and make them non-scrollable
+
 ## 4.12.2
 > Date: 2025-06-15
 ### Added